+++ CHANGES.rst
@@ -4,6 +4,14 @@ Changelog
 Pre-release (develop branch)
 ----------------------------
 
+This release requires the following new IAM permissions to function:
+
+* ``rds:DescribeAccountAttributes``
+* ``iam:GetAccountSummary``
+* ``s3:ListAllMyBuckets``
+
+Issues addressed:
+
 * `#150 <https://github.com/jantman/awslimitchecker/issues/150>`_ add CHANGES.rst to Sphinx docs
 * `#85 <https://github.com/jantman/awslimitchecker/issues/85>`_ and `#154 <https://github.com/jantman/awslimitchecker/issues/154>`_
 
@@ -13,6 +21,7 @@ Pre-release (develop branch)
     * awslimitchecker now **requires an additional IAM permission**, ``rds:DescribeAccountAttributes``
 * `#157 <https://github.com/jantman/awslimitchecker/issues/157>`_ fix for TrustedAdvisor polling multiple times - have TA set an instance variable flag when it updates services after a poll, and skip further polls and updates if the flag is set. Also add an integration test to confirm this.
 * `#50 <https://github.com/jantman/awslimitchecker/issues/50>`_ Add support for IAM service with a subset of its limits (Groups, Instance Profiles, Policies, Policy Versions In Use, Roles, Server Certificates, Users), using both limits and usage information from the `GetAccountSummary <http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetAccountSummary.html>`_ API action. This **requires an additional IAM permission**, ``iam:GetAccountSummary``.
+* `#48 <https://github.com/jantman/awslimitchecker/issues/48>`_ Add support for S3 Buckets limit. This **requires an additional IAM permission**, ``s3:ListAllMyBuckets``.
 
 0.3.2 (2016-03-11)
 ------------------